debug

+++ driver-redpanda/deploy/certification/fetch-n-report.sh
@@ -10,6 +10,10 @@ fi
 mkdir -p results
 mkdir results/$1
 ansible-playbook fetch.yaml
+if [ ! -d fetched ]; then
+  echo "failed to fetch fetched" >> log
+  exit 1
+fi
 find fetched | grep foot | xargs -I{} mv {} results/$1/footprint.tar.bz2
 rm -rf fetched
 pushd results/$1
